Move contract utils to contract-utils.ts
