+++ packages/adapter/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./chains";
 export * from "./types";
 export * from "./utils";
+export * from "./contract-utils";
 export * from "./constants";
 export * from "./errors";
